updated the logout component
